requirements.txt
aiohttp==3.8.3 ; python_version >= "3.9" and python_version < "4"
aiosignal==1.3.1 ; python_version >= "3.9" and python_version < "4"
alembic==1.9.2 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.9" and python_version < "4"
attrs==22.2.0 ; python_version >= "3.9" and python_version < "4"
base58==2.1.1 ; python_version >= "3.9" and python_version < "4"
bitarray==2.6.2 ; python_version >= "3.9" and python_version < "4"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4"
charset-normalizer==2.1.1 ; python_version >= "3.9" and python_version < "4"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
cytoolz==0.12.1 ; python_version >= "3.9" and python_version < "4"
eth-abi==2.2.0 ; python_version >= "3.9" and python_version < "4"
eth-account==0.5.9 ; python_version >= "3.9" and python_version < "4"
eth-hash==0.3.3 ; python_version >= "3.9" and python_version < "4"
eth-hash[pycryptodome]==0.3.3 ; python_version >= "3.9" and python_version < "4"
eth-keyfile==0.5.1 ; python_version >= "3.9" and python_version < "4"
eth-keys==0.3.4 ; python_version >= "3.9" and python_version < "4"
eth-rlp==0.2.1 ; python_version >= "3.9" and python_version < "4"
eth-typing==2.3.0 ; python_version >= "3.9" and python_version < "4"
eth-utils==1.10.0 ; python_version >= "3.9" and python_version < "4"
fastapi-utils==0.2.1 ; python_version >= "3.9" and python_version < "4.0"
fastapi==0.89.1 ; python_version >= "3.9" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.9" and python_version < "4"
greenlet==2.0.1 ; python_version >= "3.9" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
hexbytes==0.3.0 ; python_version >= "3.9" and python_version < "4"
idna==3.4 ; python_version >= "3.9" and python_version < "4"
ipfshttpclient==0.8.0a2 ; python_version >= "3.9" and python_version < "4"
jsonschema==4.17.3 ; python_version >= "3.9" and python_version < "4"
lru-dict==1.1.8 ; python_version >= "3.9" and python_version < "4"
mako==1.2.4 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.9" and python_version < "4.0"
multiaddr==0.0.9 ; python_version >= "3.9" and python_version < "4"
multidict==6.0.4 ; python_version >= "3.9" and python_version < "4"
netaddr==0.8.0 ; python_version >= "3.9" and python_version < "4"
parsimonious==0.8.1 ; python_version >= "3.9" and python_version < "4"
protobuf==3.19.5 ; python_version >= "3.9" and python_version < "4"
psycopg2-binary==2.9.5 ; python_version >= "3.9" and python_version < "4.0"
pycryptodome==3.16.0 ; python_version >= "3.9" and python_version < "4"
pydantic==1.10.4 ; python_version >= "3.9" and python_version < "4.0"
pyrsistent==0.19.3 ; python_version >= "3.9" and python_version < "4"
python-dotenv==0.21.1 ; python_version >= "3.9" and python_version < "4"
pywin32==305 ; python_version >= "3.9" and python_version < "4" and platform_system == "Windows"
requests==2.28.2 ; python_version >= "3.9" and python_version < "4"
rlp==2.0.1 ; python_version >= "3.9" and python_version < "4"
six==1.16.0 ; python_version >= "3.9" and python_version < "4"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
sqlalchemy==1.4.46 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.22.0 ; python_version >= "3.9" and python_version < "4.0"
toolz==0.12.0 ; python_version >= "3.9" and python_version < "4"
typing-extensions==4.4.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.9" and python_version < "4"
uvicorn==0.20.0 ; python_version >= "3.9" and python_version < "4.0"
varint==1.0.2 ; python_version >= "3.9" and python_version < "4"
web3==5.31.3 ; python_version >= "3.9" and python_version < "4"
websockets==9.1 ; python_version >= "3.9" and python_version < "4"
yarl==1.8.2 ; python_version >= "3.9" and python_version < "4"
